Filters eq, notEq, in, notIn, less, more

+++ backend/n2o/n2o-autoconfigure/src/test/java/net/n2oapp/framework/boot/mongodb/MongoDbDataProviderEngineTest.java
@@ -190,8 +190,8 @@ public class MongoDbDataProviderEngineTest {
     @Test
     public void countDocumentsOperationTest() {
         provider.setOperation(N2oMongoDbDataProvider.Operation.countDocuments);
-        Long count = (Long) engine.invoke(provider, new HashMap<>());
-        assertThat(count, is(5L));
+        Integer count = (Integer) engine.invoke(provider, new HashMap<>());
+        assertThat(count, is(5));
     }
 
     @Test
